fix: move isSessionInSample() to core
