+++ packages/session-replay-browser/src/helpers.ts
@@ -99,25 +99,6 @@ export const maskFn =
     return isMasked(elementType, config, element) ? text.replace(/[^\s]/g, '*') : text;
   };
 
-export const generateHashCode = function (str: string) {
-  let hash = 0;
-  if (str.length === 0) return hash;
-  for (let i = 0; i < str.length; i++) {
-    const chr = str.charCodeAt(i);
-    hash = (hash << 5) - hash + chr;
-    hash |= 0;
-  }
-  return hash;
-};
-
-export const isSessionInSample = function (sessionId: string | number, sampleRate: number) {
-  const hashNumber = generateHashCode(sessionId.toString());
-  const absHash = Math.abs(hashNumber);
-  const absHashMultiply = absHash * 31;
-  const mod = absHashMultiply % 1000000;
-  return mod / 1000000 < sampleRate;
-};
-
 export const getCurrentUrl = () => {
   const globalScope = getGlobalScope();
   return globalScope?.location ? globalScope.location.href : '';
